# Conflicts:
#	solution-tablet/src/com/parsroyal/solutiontablet/ui/MainActivity.java
#	solution-tablet/src/com/parsroyal/solutiontablet/ui/adapter/QuestionsAdapter.java
#	solution-tablet/src/com/parsroyal/solutiontablet/ui/fragment/dialogFragment/QuestionDetailDialogFragment.java

+++ solution-tablet/src/com/parsroyal/solutiontablet/ui/fragment/QuestionsListFragment.java
@@ -4,6 +4,7 @@ import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.RecyclerView.LayoutManager;
+import android.text.TextUtils;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,9 +15,12 @@ import butterknife.ButterKnife;
 import butterknife.Unbinder;
 import com.parsroyal.solutiontablet.R;
 import com.parsroyal.solutiontablet.constants.Constants;
+import com.parsroyal.solutiontablet.constants.PageStatus;
 import com.parsroyal.solutiontablet.data.entity.Customer;
+import com.parsroyal.solutiontablet.data.entity.Question;
 import com.parsroyal.solutiontablet.data.listmodel.QuestionListModel;
 import com.parsroyal.solutiontablet.data.listmodel.QuestionnaireListModel;
+import com.parsroyal.solutiontablet.data.model.QuestionDto;
 import com.parsroyal.solutiontablet.data.searchobject.QuestionSo;
 import com.parsroyal.solutiontablet.service.QuestionnaireService;
 import com.parsroyal.solutiontablet.service.impl.CustomerServiceImpl;
@@ -54,6 +58,7 @@ public class QuestionsListFragment extends BaseFragment {
   private int parent;
   private long answersGroupNo;
   private Customer customer;
+  private PageStatus pageStatus;
 
   public QuestionsListFragment() {
     // Required empty public constructor
@@ -85,7 +90,12 @@ public class QuestionsListFragment extends BaseFragment {
     goodsGroupBackendId = arguments.getLong(Constants.GOODS_GROUP_BACKEND_ID);
     parent = arguments.getInt(Constants.PARENT, 0);
     answersGroupNo = arguments.getLong(Constants.ANSWERS_GROUP_NO, -1);
-
+    pageStatus = (PageStatus) arguments.getSerializable(Constants.PAGE_STATUS);
+    if (pageStatus != null && pageStatus == PageStatus.VIEW) {
+      mainActivity.setToolbarIconVisibility(R.id.save_img, View.GONE);
+    } else {
+      mainActivity.setToolbarIconVisibility(R.id.save_img, View.VISIBLE);
+    }
     customer = customerService.getCustomerById(customerId);
 
     setUpRecyclerView();
@@ -98,7 +108,7 @@ public class QuestionsListFragment extends BaseFragment {
   private void setUpRecyclerView() {
     QuestionsAdapter questionsAdapter = new QuestionsAdapter(this, mainActivity, getQuestions(),
         visitId,
-        goodsGroupBackendId, answersGroupNo, customerId, questionnaireBackendId);
+        goodsGroupBackendId, answersGroupNo, customerId, questionnaireBackendId, pageStatus);
     LayoutManager layoutManager = new LinearLayoutManager(mainActivity);
     recyclerView.setLayoutManager(layoutManager);
     recyclerView.setAdapter(questionsAdapter);
@@ -128,15 +138,19 @@ public class QuestionsListFragment extends BaseFragment {
   }
 
   public void exit() {
-    DialogUtil.showCustomDialog(mainActivity, getString(R.string.warning),
-        getString(R.string.message_save_questionnaire), "", (dialog, which) ->
-            mainActivity.findViewById(R.id.save_img).performClick(),
-        "", (dialog, which) -> deleteAllQuestions(), Constants.ICON_MESSAGE);
+    if (pageStatus != null && pageStatus == PageStatus.VIEW) {
+      mainActivity.onSaveImageClicked(false);
+    } else {
+      DialogUtil.showCustomDialog(mainActivity, getString(R.string.warning),
+          getString(R.string.message_save_questionnaire), "", (dialog, which) ->
+              mainActivity.onSaveImageClicked(true),
+          "", (dialog, which) -> deleteAllQuestions(), Constants.ICON_MESSAGE);
+    }
   }
 
   private void deleteAllQuestions() {
     questionnaireService.deleteAllAnswer(visitId, answersGroupNo);
-    mainActivity.findViewById(R.id.save_img).performClick();
+    mainActivity.onSaveImageClicked(false);
   }
 
   @Override
@@ -161,10 +175,26 @@ public class QuestionsListFragment extends BaseFragment {
   }
 
   public void closeVisit() {
-    if (Empty.isEmpty(customer) || parent == MainActivity.NEW_CUSTOMER_FRAGMENT_ID) {
+    if (pageStatus == null || pageStatus == PageStatus.EDIT || Empty.isEmpty(customer)
+        || parent == MainActivity.NEW_CUSTOMER_FRAGMENT_ID) {
       //It's anonymous questionaire or New customer
       //known bug, it he has not answered any quesiton, should remove the entire visit.
       visitService.finishVisiting(visitId);
     }
+  }
+
+  public boolean hasRequiredQuestionAnswer() {
+    if (pageStatus != null && pageStatus == PageStatus.VIEW) {
+      return true;
+    }
+    for (QuestionListModel question : getQuestions()) {
+      QuestionDto questionDto = questionnaireService
+          .getQuestionDto(question.getPrimaryKey(), visitId, goodsGroupBackendId,
+              answersGroupNo);
+      if (questionDto.isRequired() && TextUtils.isEmpty(questionDto.getAnswer())) {
+        return false;
+      }
+    }
+    return true;
   }
 }
